Inferring copy number profile (CNP).

This function performs data filtering and whole genome copy number inferring.

Usage

runCNV(
  object,
  gene_order_file = NULL,
  annotations_file = NULL,
  ref_group_names,
  min_ave_counts_per_gene = 0.01,
  min_spots_per_gene = 3,
  chr_exclude = c("chrX", "chrY", "chrM"),
  max_centered_threshold = 3,
  normalization = TRUE,
  normalize_factor = NULL,
  window_length = 101,
  denoise = TRUE
)

Arguments

object

TUSCAN object.

gene_order_file

Gene location file, containing chromosome, start, and stop position for each gene.

example: chr start stop DDX11L1 chr1 11869 14412 WASH7P chr1 14363 29806 FAM138A chr1 34554 36081 OR4F5 chr1 69091 70008 OR4F29 chr1 367640 368634 OR4F16 chr1 621059 622053 ...

annotations_file

File containing annotations of all spots, with two columns representing the spot barcode and the spot clustering name (support '.txt' and '.csv' file). If the user do not provide an annotation file, the pre_cluster result will be used as the annotations.

example: V1 V2 10x13 1 10x14 2 10x15 1 10x16 1 ...

ref_group_names

A vector containing the cluster name of the reference (normal) spots used for inferring CNV.

min_ave_counts_per_gene

Minimum average gene expression level across all spots (default: 0.01).

min_spots_per_gene

Minimum number of spots requiring expression measurements to include the corresponding gene (default: 3).

chr_exclude

List of chromosomes in the reference genome annotations that should be excluded from analysis (default: c('chrX', 'chrY', 'chrM').

max_centered_threshold

Maximum value a value can have after centering. Also sets a lower bound of -1 * this value (default: 3).

normalization

Whether perform normalization step. We suggest to skip this step if the SRT data is generated from ST platform (default: TRUE).

normalize_factor

If the normalize factor is null (default setting), normalize_factor=median(UMI per spot)/UMI per spot.

window_length

Length of the window for the moving average (smoothing). Should be an odd integer (default: 101).

denoise

Whether perform de-noise step (default: TRUE).

Value

Return TUSCAN object containing inferred CNV data